Promote device interface to HalideRuntime.h and use it in HalideBuffer
removes the need for global declarations of things like
halide_copy_to_device, which conflicts with AOT code in the same
process.

This will need some more thought, because the global functions that are
no longer being called were doing useful things like error handling,
bouncing between device APIs, etc.
